% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tools_angle.R
\name{angle}
\alias{angle}
\alias{ab_angle}
\alias{cal_azimuth}
\alias{deg2rad}
\alias{rad2deg}
\title{angle functions}
\usage{
ab_angle(a, b)

angle(a, b)

cal_azimuth(x)

deg2rad(x)

rad2deg(x)
}
\description{
angle functions
}
\keyword{internal}
