% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/anusplin.R
\name{anusplin_write}
\alias{anusplin_write}
\title{ANUSPLIN writer}
\usage{
anusplin_write(
  outdir,
  dat,
  opt_splina = NULL,
  opt_lapgrd = NULL,
  na.width = "auto",
  names = c("splina.txt", "lapgrd.txt"),
  is.run = FALSE,
  ...
)
}
\arguments{
\item{outdir}{The output directory.}

\item{dat}{A formatted data.frame (data.table), or a list contains \code{dat},
\code{opt_splina} and \code{opt_lapgrd}}

\item{opt_splina}{A list of splina parameters.}

\item{opt_lapgrd}{A list of lapgrd parameters.}

\item{na.width}{The length of the whitespace to fill in the missing value,
automatic calculation (\strong{default}) or manual setting an integer.}

\item{names}{Filenames of splina and lapgrd paramters, no path required.}

\item{is.run}{If \code{TRUE}, run the command.}
}
\description{
Write the formatted data, splina and lapgrd configuration files to the same
directory.
}
