% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/weight.R
\name{cal_weight}
\alias{cal_weight}
\alias{cal_weight_sf}
\title{Weights for spatial interpolation}
\usage{
cal_weight(
  points,
  range,
  res = 0.5,
  cdd = 450,
  m = 4,
  nstation.max = 8,
  nstation.min = 3,
  wFUN = c("wFUN_adw", "wFUN_idw"),
  .progress = "none",
  ...
)

cal_weight_sf(
  points,
  range = NULL,
  res = 0.25,
  cdd = 450,
  m = 4,
  nstation.max = 8,
  nstation.min = 3,
  wFUN = c("wFUN_adw", "wFUN_idw"),
  .progress = "none",
  ...
)
}
\arguments{
\item{points}{A matrix (N,2) with longitude and latitude of points of data observed}

\item{range}{\verb{[xmin, xmax, ymin, ymax]}}

\item{res}{the grid resolution (degree)}

\item{cdd}{the correlation decay distance (km) for \code{adw} and searching radius
for others, default (450).}

\item{m}{distance weight (default 4)}

\item{nstation.max}{Number of maximum stations used per point for
interpolation, (default 8).}

\item{nstation.min}{Number of minimum stations used per point for
interpolation, (default 3).}

\item{.progress}{name of the progress bar to use, see
\code{\link[plyr]{create_progress_bar}}}

\item{...}{other parameters to \link[plyr:ldply]{plyr::ldply}}
}
\value{
A data.frame with longitude, latitude and interpoled points
}
\description{
Weights for spatial interpolation
}
