% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/kfold.R
\name{kford_ml}
\alias{kford_ml}
\title{kfold machine learning}
\usage{
kford_ml(
  X,
  Y,
  kfold = 5,
  FUN,
  ...,
  weight = NULL,
  Z = NULL,
  seed = 1,
  na.rm = FALSE
)
}
\description{
kfold machine learning
}
