% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/anusplin_helper.R
\name{query_grid_type}
\alias{query_grid_type}
\title{query_grid_type}
\usage{
query_grid_type(name = "arcinfo_grid")
}
\description{
query_grid_type
}
\section{grid_type}{

\itemize{
\item \code{none}         : not provided
\item \code{xyz}          : grid written in X,Y,Z format
\item \code{generic_grid} : generic grid written by rows
\item \code{arcinfo_grid} : Arc/Info grid
\item \code{idrisi_img}   : Idrisi image
}
}

