% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/spInterp-package.R
\docType{package}
\name{spInterp-package}
\alias{spInterp}
\alias{spInterp-package}
\title{spInterp: A toolbox for spatial interpolation of meteorogical data}
\description{
Angular Distance Weighting and Inverse Distance Weighting methods for interpolation of irregular network of meteorological observations to a regular grid.
}
\author{
\strong{Maintainer}: Dongdong Kong \email{kongdd.sysu@gmail.com} (\href{https://orcid.org/0000-0003-1836-8172}{ORCID})

Authors:
\itemize{
  \item Heyang Song \email{7800556@gmail.com} (\href{https://orcid.org/0000-0002-4192-5603}{ORCID})
}

}
\keyword{internal}
