% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/spInterp_anusplin.R
\name{spInterp_anusplin}
\alias{spInterp_anusplin}
\title{spInterp_anusplin}
\usage{
spInterp_anusplin(
  X,
  Y,
  range,
  res = 1,
  outdir = "output",
  prefix = "RH",
  file.alt = NULL,
  overwrite = TRUE,
  ...,
  Z = NULL
)
}
\arguments{
\item{...}{others to \code{\link[=anusplin_make_param]{anusplin_make_param()}}}
}
\description{
spInterp_anusplin
}
