% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/spInterp_adw.R
\name{weight_adw}
\alias{weight_adw}
\alias{weight_adw_sf}
\title{Weights of Angular Distance Weighting interpolation}
\usage{
weight_adw(
  points,
  range,
  res = 0.5,
  cdd = 450,
  m = 4,
  nstation.max = 8,
  nstation.min = 3,
  .progress = "text",
  ...
)

weight_adw_sf(
  points,
  range = NULL,
  res = 0.25,
  cdd = 450,
  m = 4,
  nstation.max = 8,
  nstation.min = 3,
  .progress = "text",
  ...
)
}
\arguments{
\item{points}{A matrix (N,2) with longitude and latitude of points of data observed}

\item{range}{`[xmin, xmax, ymin, ymax]`}

\item{res}{the grid resolution}

\item{cdd}{the correlation decay distance (km), default (450)}

\item{m}{distance weight (default 4)}

\item{...}{other parameters to [plyr::ldply]}

\item{n.station}{Number of stations used per point for interpolation. The default is 8.}
}
\value{
A data.frame with longitude, latitude and interpoled points
}
\description{
The irregularly-spaced data are interpolated onto regular latitude-longitude
grids by weighting each station according to its distance and angle from the
center of a search radius.
}
\examples{
library(ggplot2)

data(TempBrazil) # Temperature for some poins of Brazil

loc <- TempBrazil[, 1:2] \%>\% set_names(c("lon", "lat"))
Temp <- TempBrazil[, 3] # Vector with observations in points

range <- c(-78, -34, -36, 5)

# 
ggplot(r, aes(lon, lat)) +
  geom_raster(aes(fill = value)) +
  geom_point(data = loc, size = 2.5, shape = 3, color = "red")

profvis::profvis({
# proffer::pprof({
# system.time({
  # r <- ADW(xy = loc, z = Temp, xrange = range[1:2], yrange = range[3:4])
  # l <- weight_adw(loc, range = range, res = 2, .progress = "text")
  r <- weight_adw(loc, range = range, res = 1)
})
x = rnorm(100)

system.time({
  r1 = weight_adw(loc, range, res = 0.25)
  # r2 = weight_adw_sf(loc, range, res = 2)
})


compare(
  sapply(r1, nrow) ,
  sapply(r2, nrow)  
)

system.time(
  foreach(i = 1:1e3) \%do\% {
    cal_hw(c(100, 20))
  }
)

dat = matrix(rnorm(265*10), 265)

# 112
res = spInterp_adw(loc, dat)


set.seed(2)
dd <- data.frame(
  lon = runif(100, min = 110, max = 117),
  lat = runif(100, min = 31, max = 37),
  value = runif(100, min = -10, max = 10)
)
head(dd)
dg <- adw(dd, gridSize = 1, cdd = 1e5)
# dg is the dataframe of grid (mesh)
head(dg)

}
\references{
1. Xavier, A. C., King, C. W., & Scanlon, B. R. (2016). Daily gridded 
   meteorological variables in Brazil (1980–2013). International Journal of 
   Climatology, 36(6), 2644–2659. <doi:10.1002/joc.4518>
}
\author{
Dongdong Kong and Heyang Song
}
