% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wFUN.R
\name{wFUN}
\alias{wFUN}
\alias{wFUN_adw}
\alias{wFUN_idw}
\alias{wFUN_mean}
\alias{wFUN_thiessen}
\alias{wFUN_natural}
\title{function to calculate sites weights}
\usage{
wFUN_adw(dist, m = 4, cdd = 450, point, sites, ...)

wFUN_idw(dist, m = 2, ...)

wFUN_mean(dist, ...)

wFUN_thiessen(dist, ...)

wFUN_natural(dist, ...)
}
\arguments{
\item{dist}{numeric vector, the distance between point and each sites, with the
same length as \code{sites}.}

\item{m}{distance weight (default 4)}

\item{cdd}{the correlation decay distance (km) for \code{adw} and searching radius
for others, default (450).}

\item{point}{numeric vector \code{c(lon, lat)}, the point needed to be interpolated}

\item{sites}{matrix, \code{cbind(lon, lat)}, coordinates of sites used to interpolate \code{point}}

\item{...}{ignored}
}
\description{
\itemize{
\item \code{wFUN_adw}: angular distance weighting
\item \code{wFUN_idw}: inverse distance weighting
\item \code{wFUN_mean}: arithmetic average weighting
\item \code{wFUN_thiessen},\code{wFUN_natural}: natural neighbor interpolation (also known
as thiessen interpolation)
}
}
