% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/write_anusplin.R
\name{anusplin_write}
\alias{anusplin_write}
\title{ANUSPLIN writer}
\usage{
anusplin_write(
  dat,
  opt_splina,
  opt_lapgrd,
  file_path,
  exe,
  na.width = "auto",
  names = c("splina.txt", "lapgrd.txt"),
  cmd = NULL
)
}
\arguments{
\item{dat}{A formatted data.frame (data.table).}

\item{opt_splina}{A vector of splina parameters.}

\item{opt_lapgrd}{A vector of lapgrd parameters.}

\item{file_path}{The output directory.}

\item{exe}{Path to splina and lapgrd exe.}

\item{na.width}{The length of the whitespace to fill in the missing value,
automatic calculation (\strong{default}) or manual setting an integer.}

\item{names}{Filenames of splina and lapgrd paramters, no path required.}

\item{cmd}{The filename of batch execution script, or no export (\strong{default}).}
}
\description{
Write the formatted data, splina and lapgrd configuration files to the same
directory.
}
