% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/reexports.R, R/tools_Ipaper.R
\docType{import}
\name{reexports}
\alias{reexports}
\alias{\%>\%}
\alias{\%<>\%}
\alias{\%$\%}
\alias{set_names}
\alias{Tps}
\alias{predictSurface}
\alias{image.plot}
\alias{rowMeans2}
\title{Objects exported from other packages}
\keyword{internal}
\description{
These objects are imported from other packages. Follow the links
below to see their documentation.

\describe{
  \item{fields}{\code{\link[fields]{image.plot}}, \code{\link[fields]{predictSurface}}, \code{\link[fields]{Tps}}}

  \item{magrittr}{\code{\link[magrittr:exposition]{\%$\%}}, \code{\link[magrittr:compound]{\%<>\%}}, \code{\link[magrittr:pipe]{\%>\%}}, \code{\link[magrittr:aliases]{set_names}}}

  \item{matrixStats}{\code{\link[matrixStats]{rowMeans2}}}
}}

