% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/spInterp-package.R
\docType{package}
\name{spInterp-package}
\alias{spInterp}
\alias{spInterp-package}
\title{spInterp: A toolbox for spatial interpolation of meteorogical data}
\description{
Angular Distance Weighting and Inverse Distance Weighting methods for interpolation of irregular network of meteorological observations to a regular grid.
}
\author{
\strong{Maintainer}: Rodrigo Lins Júnior \email{rodrigo.lins@simepar.br}

}
\keyword{internal}
